update ngày 19/09/2017

+++ web/src/main/java/com/cmms/webapp/action/MachineTypeAction.java
@@ -141,13 +141,26 @@ public class MachineTypeAction extends BaseAction implements Preparable {
         try {
             JSONObject result = new JSONObject();
             if (ids != null && ids.length > 0) {
+                List<Integer> list = new ArrayList<>(ids.length);
                 if (ids.length == 1) {
+                    list.add(Integer.parseInt(ids[0]));
+                    if (machineTypeDao.checkUseMechanic(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
                     machineTypeDao.remove(Integer.parseInt(ids[0]));
                 } else {
-                    List<Integer> list = new ArrayList<>(ids.length);
                     for (String idTmp : ids) {
                         list.add(Integer.parseInt(idTmp));
                     }
+
+                    if (machineTypeDao.checkUseMechanic(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
+                    
                     int delete = machineTypeDao.delete(list);
                     if (delete != ids.length) {
                         log.warn("deleteCompany rtn: " + delete + " list: " + ids.length);
