+++ web/src/main/java/com/cmms/webapp/action/WorkTypeAction.java
@@ -143,12 +143,24 @@ public class WorkTypeAction extends BaseAction implements Preparable {
         try {
             JSONObject result = new JSONObject();
             if (ids != null && ids.length > 0) {
+                List<Integer> list = new ArrayList<>(ids.length);
                 if (ids.length == 1) {
+                    list.add(Integer.parseInt(ids[0]));
+                    if (workTypeDao.checkUseWO(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
                     workTypeDao.remove(Integer.parseInt(ids[0]));
                 } else {
-                    List<Integer> list = new ArrayList<>(ids.length);
+                    list = new ArrayList<>(ids.length);
                     for (String idTmp : ids) {
                         list.add(Integer.parseInt(idTmp));
+                    }
+                    if (workTypeDao.checkUseWO(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
                     }
                     int delete = workTypeDao.delete(list);
                     if (delete != ids.length) {
